Resolving issues with cache
