+++ lib/src/cache/byte_storage_abstract.dart
@@ -1,0 +1,9 @@
+import 'dart:typed_data';
+
+abstract class AbstractByteStorage<TDirectory, TFile> {
+  Future<void> write(String path, Uint8List bytes);
+  Future<Uint8List?> read(String path);
+  Future<void> delete(String path);
+  Future<TDirectory?> storageDirectory();
+  Future<TFile?> fileOf(String path);
+}
