+++ pubspec.yaml
@@ -18,6 +18,7 @@ dependencies:
     #path: ../vector_tile_renderer
   async: ^2.8.2
   executor_lib: ^1.1.1
+  idb_shim: ^2.6.1+7
     #path: ../executor_lib
 
 dev_dependencies:
